mcp-echo-streamablehttp-server-stateless/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-echo-streamablehttp-server-stateless"
version = "0.1.0"
description = "A stateless MCP server implementing echo and printEnv tools via streamablehttp transport"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "MCP Echo Server", email = "noreply@example.com" }
]
dependencies = [
    "mcp>=1.3.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
mcp-echo-streamablehttp-server-stateless = "mcp_echo_streamablehttp_server_stateless.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_echo_streamablehttp_server_stateless"]